Fixes running on newer versions of Perl
Specifically fixed a "uniq" function that returned the unique
entries in an array. The cryptic function was using an
experimental "keys on scalar" feature that is now forbidden
in newer versions of perl.

I replaced the uniq function with something less crtypic
and that works on older and newer versions of perl.

Tested on Perl 5.32.1 and 5.18.4